% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMD.R
\docType{package}
\name{clustMD-package}
\alias{_PACKAGE}
\alias{clustMD-package}
\title{Model based clustering for mixed data: clustMD}
\description{
Model-based clustering of mixed data (i.e. data that consist of continuous,
binary, ordinal or nominal variables) using a parsimonious mixture of latent
Gaussian variable models.
}
\author{
Damien McParland

Damien McParland <damien.mcparland@ucd.ie>
Isobel Claire Gormley <claire.gormley@ucd.ie>
}
\references{
McParland, D. and Gormley, I.C. (2016). Model based clustering 
    for mixed data: clustMD. Advances in Data Analysis and Classification, 
    10 (2):155-169.
}
\seealso{
\code{\link{clustMD}}
}
\keyword{package}

