% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMD_UtilityFunctions.R
\name{getOutput_clustMDparallel}
\alias{getOutput_clustMDparallel}
\title{Extracts relevant output from \code{clustMDparallel} object}
\usage{
getOutput_clustMDparallel(resParallel, nClus, covModel)
}
\arguments{
\item{resParallel}{a \code{clustMDparallel} object.}

\item{nClus}{the number of clusters in the desired output.}

\item{covModel}{the covariance model of the desired output.}
}
\value{
A \code{clustMD} object containing the output for the relevant 
    model.
}
\description{
This function takes a \code{clustMDparallel} object, a number of clusters
and a covariance model as inputs. It then returns the output corresponding
to that model. If the particular model is not contained in the 
\code{clustMDparallel} object then the function returns an error.
}
