% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMD_UtilityFunctions.R
\name{modal.value}
\alias{modal.value}
\title{Calculate the mode of a sample}
\usage{
modal.value(x)
}
\arguments{
\item{x}{a vector containing the sample values.}
}
\value{
The mode of the sample. In the case of a tie, the minimum is 
    returned.
}
\description{
Calculate the mode of a sample
}
\keyword{internal}
