% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMD_UtilityFunctions.R
\name{patt.equal}
\alias{patt.equal}
\title{Check if response patterns are equal}
\usage{
patt.equal(x, patt)
}
\arguments{
\item{x}{a numeric vector.}

\item{patt}{a vector to compare \code{x} to.}
}
\value{
Returns \code{TRUE} if \code{x} and \code{patt} are exactly the same
    and \code{FALSE} otherwise.
}
\description{
Checks whether response patterns are equal or not and returns \code{TRUE}
or \code{FALSE} reprectively.
}
\note{
Used internally in clustMD function.
}
\keyword{internal}
