% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMD_Methods.R
\name{plot.clustMDparallel}
\alias{plot.clustMDparallel}
\title{Summary plots for a clustMDparallel object}
\usage{
\method{plot}{clustMDparallel}(x, ...)
}
\arguments{
\item{x}{a \code{clustMDparallel} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Produces a number of plots as detailed above.
}
\description{
Produces a line plot of the estimated BIC values corresponding to each 
covariance model against the number of clusters fitted. For the optimal model
according to this criteria, a parallel coordinates plot of the cluster means
is produced along with a barchart or heatmap of the covariance matrices for
each cluster and a histogram of the clustering uncertainties.
}
