% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMD_Methods.R
\name{print.clustMD}
\alias{print.clustMD}
\title{Print basic details of \code{clustMD} object.}
\usage{
\method{print}{clustMD}(x, ...)
}
\arguments{
\item{x}{a \code{clustMD} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints summary details, as described above, to screen.
}
\description{
Prints a short summary of a \code{clustMD} object to screen. Details the
number of clusters fitted as well as the covariance model and the estimated
BIC.
}
\seealso{
\code{\link{clustMD}}
}
\keyword{print}
