% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMD_Methods.R
\name{print.clustMDparallel}
\alias{print.clustMDparallel}
\title{Print basic details of \code{clustMDparallel} object}
\usage{
\method{print}{clustMDparallel}(x, ...)
}
\arguments{
\item{x}{a \code{clustMDparallel} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints details described above to screen.
}
\description{
Prints basic details of \code{clustMDparallel} object. Outputs the different
numbers of clusters and the different covariance structures fitted to the
data. It also states which model was optimal according to the estimated BIC
criterion.
}
\seealso{
\code{\link{clustMD}}
}
\keyword{print}
