% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMD_UtilityFunctions.R
\name{qfun}
\alias{qfun}
\title{Helper internal function for \code{dtmvnom()}}
\usage{
qfun(a, b, S)
}
\arguments{
\item{a}{a vector of lower thresholds.}

\item{b}{a vector of upper thresholds.}

\item{S}{the covariance matrix of the untruncated distribution.}
}
\value{
Output required for dtmvnom function.
}
\description{
Internal function.
}
\references{
Kan, R., & Robotti, C. (2016). On Moments of Folded and 
    Truncated Multivariate  Normal Distributions. Available at SSRN.
}
\keyword{distribution}
\keyword{internal}
