% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMD_Methods.R
\name{summary.clustMD}
\alias{summary.clustMD}
\title{Summarise \code{clustMD} object}
\usage{
\method{summary}{clustMD}(object, ...)
}
\arguments{
\item{object}{a \code{clustMD} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints summary of \code{clustMD} object to screen, as detailed above.
}
\description{
Prints a summary of a \code{clustMD} object to screen. Details the number
of clusters fitted as well as the covariance model and the estimated BIC.
Also prints a table detailing the number of observations in each cluster and
a matrix of the cluster means.
}
\seealso{
\code{\link{clustMD}}
}
\keyword{print}
