% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMD_Methods.R
\name{summary.clustMDparallel}
\alias{summary.clustMDparallel}
\title{Prints a summary of a clustMDparallel object to screen.}
\usage{
\method{summary}{clustMDparallel}(object, ...)
}
\arguments{
\item{object}{a \code{clustMDparallel} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints a summary of the \code{clustMDparallel} object to screen, as
    detailed above.
}
\description{
Prints the different numbers of clusters and covariance models fitted and
indicates the optimal model according to the estimated BIC criterion. The
estimated BIC for the optimal model is printed to screen along with a table
of the cluster membership and the matrix of cluster means for this optimal
model.
}
\seealso{
\code{\link{clustMD}}
}
\keyword{print}
