% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMD_UtilityFunctions.R
\name{vec.outer}
\alias{vec.outer}
\title{Calculate the outer product of a vector with itself}
\usage{
vec.outer(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Returns the outer product of the vector \code{x} with itself.
}
\description{
This function calculates the outer product of a vector with itself.
}
\keyword{internal}
