\name{z.nom.diag}
\alias{z.nom.diag}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Transforms Monte Carlo simulated data into categorical data. Calculates empirical moments of latent data given categorical responses.
}
\description{
Internal function.
}
\usage{
z.nom.diag(Z)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Z}{
	A matrix of Monte Carlo simulated data.
}
}
\details{
z.nom.diag: an internal function.
}
\value{
	Output required for clustMD function.
}
\references{
	McParland, D. and Gormley, I.C. (2014). Model based clustering for mixed data: clustMD. Technical report, University College Dublin.
}
\author{
	Damien McParland
}
\note{
	An internal function.
}

\seealso{
	\code{\link{clustMD}}
}

\keyword{ internal }

