\name{clustMD-package}
\alias{clustMD-package}
\docType{package}
\title{
Model based clustering for mixed data
}
\description{
Model-based clustering of mixed data (i.e. data that consist of continuous, binary, ordinal or nominal variables) using a parsimonious mixture of latent Gaussian variable models.
}
\details{
\tabular{ll}{
Package: \tab clustMD\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-30\cr
License: \tab GPL-2\cr
}

}
\author{
Damien McParland 

Damien McParland <damien.mcparland@ucd.ie>
Isobel Claire Gormley <claire.gormley@ucd.ie>
}
\references{
McParland, D. and Gormley, I.C. (2014). Model based clustering for mixed data: clustMD. Technical report, University College Dublin.
}

\keyword{ package }
