% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMD_InternalFunctions.R
\name{npars_clustMD}
\alias{npars_clustMD}
\title{Calculates the number of free parameters for the \code{clustMD} model.}
\usage{
npars_clustMD(model, D, G, J, CnsIndx, OrdIndx, K)
}
\arguments{
\item{model}{the \code{clustMD} covariance model fitted.}

\item{D}{the dimension of the latent data.}

\item{G}{the number of mixture components.}

\item{J}{the number of variables.}

\item{CnsIndx}{the number of continuous variables.}

\item{OrdIndx}{the sum of the number of continuous and ordinal (including
binary) variables.}

\item{K}{a vector indicating the number of levels of each categorical 
variable.}
}
\value{
Output required for \code{clustMD} function.
}
\description{
Internal function.
}
\seealso{
\code{\link{clustMD}}
}
\keyword{internal}
