% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMD_InternalFunctions.R
\name{perc.cutoffs}
\alias{perc.cutoffs}
\title{Calculates the threshold parameters for ordinal variables.}
\usage{
perc.cutoffs(CnsIndx, OrdIndx, Y, N)
}
\arguments{
\item{CnsIndx}{the number of continuous variables.}

\item{OrdIndx}{the sum of the number of continuous and ordinal (including
binary) variables.}

\item{Y}{an \code{N x J} data matrix.}

\item{N}{number of observations.}
}
\value{
Output required for \code{clustMD} function.
}
\description{
Calculates the threshold parameters for ordinal variables.
}
\seealso{
\code{\link{clustMD}}
}
\keyword{internal}
