% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMD_InternalFunctions.R
\name{z.nom.diag}
\alias{z.nom.diag}
\title{Transforms Monte Carlo simulated data into categorical data. Calculates
empirical moments of latent data given categorical responses.}
\usage{
z.nom.diag(Z)
}
\arguments{
\item{Z}{a matrix of Monte Carlo simulated data.}
}
\value{
Output required for \code{clustMD} function.
}
\description{
Internal function.
}
\seealso{
\code{\link{clustMD}}
}
\keyword{internal}
